--- src/pages/home/components/JWTPage.style.ts
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const JWTPageStyle = styled.div`
-  display: flex;
-  width: 100%;
-  flex-direction: row;
-  gap: 40px;
-  padding: 40px;
-`;
-
-export const Wrap = styled.div`
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  gap: 40px;
-`;
