--- src/pages/home/components/Page.style.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const PageStyle = styled.div`
-  display: flex;
-  width: 100%;
-  height: 100%;
-  flex-direction: row;
-  gap: 40px;
-  padding: 40px;
-`;
