--- src/pages/home/components/Page.tsx
@@ -1,12 +1,0 @@
-import HomeRanking from './home-ranking/HomeRanking';
-import HomeStudyRooms from './home-studyrooms/HomeStudyRooms';
-import { PageStyle } from './Page.style';
-
-export default function Page() {
-  return (
-    <PageStyle>
-      <HomeStudyRooms />
-      <HomeRanking />
-    </PageStyle>
-  );
-}
